change location for common function deletedircontents to utility package, shared amongst cclf and suppression packages
